fix: address PR #1154 review on schema-version typing and atomicity JavaDoc
- Narrow EvitaSession#startCatalogSchemaVersion from long to int to match
  the actual return type of CatalogSchema#version() (pre-existing
  inconsistency, no behavioral change).
- Soften ExpandedEngineState#replaceCatalogReference JavaDoc: the prior
  schema-version read happens BEFORE the atomic swap, not inside it; the
  commit pipeline (TransactionManager#propagateCatalogSnapshot) serializes
  calls per catalog so the race is not reachable in practice.

Ref: #1153

+++ evita_engine/src/main/java/io/evitadb/core/session/EvitaSession.java
@@ -205,7 +205,7 @@ public final class EvitaSession implements EvitaInternalSessionContract {
 	 * Represents the starting version of the catalog schema.
 	 * This variable is used to keep track of the initial version of the catalog schema.
 	 */
-	private final long startCatalogSchemaVersion;
+	private final int startCatalogSchemaVersion;
 	/**
 	 * Contains reference to the proxy factory that is used to create proxies for the entities.
 	 */
